Add machines complete
There is a robust GUI now for adding RecyclingMachines to the
RecyclingStation. I've added a few more stub functions to the
RecyclingMachine class to send and retrieve settings between the Station
and the Machine.

I've designed the add machine GUI in such a way where it should be
relatively easy to implement the modify machine GUI. It can reuse the
ModifyMachinePanel which is in its own file/class.